pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ha-habitus"
version = "0.13.0"
description = "Behavioral intelligence add-on for Home Assistant — local-first, privacy-preserving"
authors = [{ name = "Craig Allen", email = "craigallen12345@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "scikit-learn>=1.4",
    "pandas>=2.1",
    "numpy>=1.26",
    "requests>=2.31",
    "websockets>=12.0",
    "flask>=3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.3",
    "black>=24.0",
    "mypy>=1.9",
    "responses>=0.25",
    "freezegun>=1.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "TCH"]
ignore = ["E501", "B008", "N803", "N806"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "ANN"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=habitus/habitus --cov-report=xml --cov-report=term-missing --cov-fail-under=70"
log_cli = true
log_cli_level = "INFO"

[tool.coverage.run]
source = ["habitus/habitus"]
omit = ["tests/*", "*/web.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
